+++ programs/stake_api/src/stake_instruction.rs
@@ -11,7 +11,8 @@ use solana_sdk::{
     instruction::{AccountMeta, Instruction, InstructionError},
     instruction_processor_utils::DecodeError,
     pubkey::Pubkey,
-    rent_utils, system_instruction, sysvar,
+    system_instruction, sysvar,
+    sysvar::rent,
 };
 
 /// Reasons the stake might have had an error
