+++ sdk/src/lib.rs
@@ -15,7 +15,6 @@ pub mod packet;
 pub mod poh_config;
 pub mod pubkey;
 pub mod rent_calculator;
-pub mod rent_utils;
 pub mod rpc_port;
 pub mod short_vec;
 pub mod system_instruction;
